chore(release): v26.5.5 — Claude Code Maximization & Development Infrastructure
This release activates bleeding-edge Claude Code features and completes development infrastructure hardening:

**Claude Code Configuration**
- Wired 5 essential hooks (SessionStart, UserPromptSubmit, Stop, PreToolUse, PostToolUse)
- Enabled automated evidence collection and release gate enforcement
- Created cargo-make-runner, otel-span-verifier, and andon-stop skills

**Skills Infrastructure**
- Added YAML frontmatter path scoping to context-aware skills
- Reduced cognitive load by loading only relevant skills per file context

**Bug Fixes**
- Test mutability warnings resolved (ggen-core, ggen-cli, prolog8)
- Import resolution corrected in template inheritance and simple_cycle_test
- Enhanced domain utilities with improved error handling

Closes: Manufacturing discipline hardening (STOP the line, OTEL validation, cargo-make enforcement)
See: .claude/IMPLEMENTATION_SUMMARY.md for detailed changes

BREAKING: None
Performance: No change

+++ crates/ggen-core/tests/rendering_property_tests.rs
@@ -198,7 +198,7 @@ proptest! {
             .filter(|l| l.trim().starts_with("// -"))
             .count();
 
-        prop_assert_eq!(tool_lines, num_tools as usize);
+        prop_assert_eq!(tool_lines, num_tools);
     }
 }
